Fixed terrible bugs introduced in the previous patch.
Most hovering panels can no longer match.
Panels that hover because a clear finished under them can match.
Panels that hover because they finished a swap on top of a panel that is
hovering because a clear finished under it can match.

It's obvious that this will leave the desired behavior in place where
swapping things onto matches as they finish is concerned, but I am not
certain that everything else will work correctly yet...