fixed crash for chains >4 and fixed last pop in a combo not playing
had forgotten before this to not try to play a pop with chain counters above 4, as there aren't any different sound effects for that.  Now it will just play level 4 sound effects for larger chains.

Also added code to play a pop for the last panel in a combo to pop.