saving progress.  still a bug
there is still a bug with the server not making the correct decision as to who won on games after the first game in a session.  I think it has to do with the room.game_outcome_reports not getting reset to nil properly when a new game starts.  I think after the first match it is only considering the outcome report from the first player who reports in, and uses the other players previous report.