include in the leaderboard report whether player is the player that requested the report.
planning to show on the client's leaderboard screen which player on the list belongs to the client.  Adding another field is_you to the leaderboard report.  it will be true for the line item in the list where the user_id_of_requester passed in matches the user_id of the player getting inserted into the report.