initial cursors
We'll want to put the padding in load_img back when @sharpobject writes a wrapper for getDimensions

TO DO: use cursor halves on the edges of wide buttons (or all of them), instead of stretching the cursor.

draw the cursor on the outside of the buttons, instead of in them.