improved text appearance on canvas
still not perfect, it seems, but definitely not as blurry when the window is resized.  Before this pull request, when you'd resize, text really got blurry.  This commit should improve that, but I don't know why it's blurry at all at default window size.  It didn't seem to be on love 0.9.0.