Fix garbage panel hover timing
Newly-transformed panels from garbage should begin hovering immediately
after the last bit of garbage is done popping, and should hover for
different amounts of time based on the current game level.  Adds a table
of timing values, and changes garbage animation timing so it doesn't
count the "bounce" panels do after garbage is finished transforming as
part of the transforming animation.

A consequence of this change is that the "bounce" is now timed
incorrectly (it plays while garbage is still transforming, instead of
afterwards).  This might need a separate flag so it can be drawn
properly?