Fix telegraph chaining off by 1 error (#444)
It's okay for chaining to be sent to the opponent on the frame it happened just like attacks, because it doesn't interact with the opponents stack on that frame.

However we still want to rollback to the frame "after" the attack or chain end because:
A) We don't need to rollback any earlier
B) We don't want to rollback to the same frame as the player thats "behind" because we may have already sent an attack that frame that the behind player processed. We only clear attacks in the "future" so we need to make sure we don't rollback to a frame thats already processed attacks. Luckily we don't need to rollback that far since the attack only interacts with the next frame at a minimum.

Also added some debugging and reordered the ifs to make more sense