Break up server into multiple object oriented files (#786)
The server files are currently thrown in one big file and all the data access is interconnected and confusing. Not only does this make it hard to follow, it makes it hard to change and hard to test.

To start I have moved all the files into a server folder and broken them up into separate files.
There is still a lot of global variable access, but I did fix any small wins I could while I was here.