Fix challenge mode absolute bar to support time label
The challenge mode absolute bar didn't show the amount of time remaining on the label like it does in vs.
Fixed.

Should the label show the actual time remaining instead of the amount remaining before the bar starts draining? To me it felt like the label should be total time.