Fix tests
This fixes all tests to be more accurate to a real world scenario where matches happen one at at time on both players before the player table is updated.
It also fixes that bug along with the bug of not updating the RD after the last game a player played.
As far as I know all feature work is done and tested well.