Fix Bool Selector for mods, plus graphics update
Fixes #597 so the selector works for both touch and buttons
Also removed duplicate code for focusable selectors by moving into BoolSelector class
Also fixed the BoolSelector graphics to make it more obvious when it is on and able to be horizontal
Used a horizontal version everywhere for now as its more standard.
Made the ranked selector have a title rather than the up down titles as that was confusing and its more obvious now.
Standarization of the focusable require name, maybe it should be named "makeFocusable"?
Fixed 2p only having border on ranked