Add a stop danger animation
Adds an animation to panels for when you are in danger but have stop time.

This allows you an easier time of telling if you have stop time rather than looking at the colors on the multibar.

This feature is optional, by default the stop danger animation is the same as the normal danger animation. Any panels that have the feature can easily be disabled by changing the row in the config.