Create Puzzle Previews
Made a preview of the panel board in the puzzle menu
Cleaned up PanelBoardElement to do transforms right. It currently can function as a UIElement or be drawn directly
Cleaned up player stack so it has helper functions for setting up board rendering
Fixed portrait game? Hard to know what was working before on desktop with different orientations and scales
Added a way for menu items to have a callback when selected