Fix Level Data too not use too precise floats
Our JSON library only sends 14 digits of precision, so we need to make sure to not store more than that
Added extensive testing and validation that we don't anymore after moving over the level data.
In practice this didn't really matter in the engine, but did cause the server to think we had different level data
Note that we will need to update the server, but we can do so without forcing stable to move immediately because it will just have the old behavior of levels 2-7 not playing ranked.