Fix Canvas Images going invisible
We were rendering to canvases and then expecting them to stick around,
but love documents canvas being cleared when the window size changes.

Instead we should render them to an image and just use the image instead.

Added helper methods that make sure to reset the graphics state when rendering

Also used the right DPI and filter.

Tested stage and character bundles