Fix Connection and TCPClient bugs with partial send
When socket sends data and it only partially sends, we get an error and the amount of data that was sent.

We are only supposed to send the rest of the data next time the socket is free but we were sending the whole thing in the Connection server case, and dropping it on the floor in the client case.

Note that we didn't have this problem on the receiving side.

Added tests to prove the problems, then fixed them.

Also note I could refactor this more, but didn't want to introduce more risk.